refactor(LC001): split local-method analysis helpers
- separate queryable-context analysis and translation trust checks
- keep analyzer behavior unchanged while reducing local complexity
